PYTHON/requirements_ga4_dashboard.txt
# Pour ga4_dashboard_report.py (API Data GA4 → HTML + Chart.js)
google-analytics-data>=0.18.0
google-auth>=2.0.0
